requirements.txt
Flask
pycryptodome
Pillow
python-dotenv
